[FIX] website: move _load_module_terms to ir.module.module
