+++ addons/website/models/__init__.py
@@ -18,7 +18,6 @@ from . import website_menu
 from . import website_page
 from . import website_rewrite
 from . import ir_rule
-from . import ir_translation
 from . import ir_ui_view
 from . import res_company
 from . import res_partner
